优化： 修改memory的length 和size 为int 并且增加小于零的判断。 修订tcp的发送逻辑。 微调tcp框架的处理方式：链接断开由处理线程触发，并发送消息包给主线程，最终由主线程来处理对应逻辑。

+++ src/fr_public/pub_memory.h
@@ -31,16 +31,19 @@ namespace Universal{
 	//				* 该类会独立的管理内存，从外界复制的数据为深复制。所以，频繁的构建该类会有大量内存的malloc和free。
 	//! \note	注意点和缺陷：
 	//				* 没有提供内存修改的函数体系。如修改数据需要直接调用指针。然而直接调用指针无法对增删代码进行响应。
+	//				* 最开始size 和 length 使用int64(即无符号整形)，但最后在使用中发现，使用int64时，碰见int的负数，会导致一些非期望行为.所以将所有int64修改为int
+	//					* 例如：length = -1的del操作，会删除所有的内存。而length=-1可能在调用本意应该是不删除任何数据(无论是输错还是如何，-1等不应该delete any data。)
+	//					* 这里注意uint 和 int之间的转换在g++某些版本的默认配置下，是没有任何提示(包括警告也是没有的)的
 	//! \todo	考虑重载 <<。接受各类基础类型转换为字节流的方式。
 	//! \todo	获取外界new指针，避免重复的内存new/malloc
 	class BinaryMemory{
 		public:
 			BinaryMemory();
-			BinaryMemory(const void *_buffer, size_t _size);
-			BinaryMemory(const Byte *_buffer, size_t _size);
-			BinaryMemory(const char *_buffer, size_t _size);
-			BinaryMemory(const uint16 *_buffer, size_t _size);
-			BinaryMemory(const uint32 *_buffer, size_t _size);
+			BinaryMemory(const void *_buffer, int64 _size);
+			BinaryMemory(const Byte *_buffer, int64 _size);
+			BinaryMemory(const char *_buffer, int64 _size);
+			BinaryMemory(const uint16 *_buffer, int64 _size);
+			BinaryMemory(const uint64 *_buffer, int64 _size);
 			BinaryMemory(const BinaryMemory &ref);
 			~BinaryMemory();
 		public:
@@ -49,21 +52,21 @@ namespace Universal{
 		public:
 			void addBuffer(const BinaryMemory &ref);
 			inline void add(const BinaryMemory &ref){ addBuffer(ref); }
-			void addBuffer(const void* buffer, size_t size);
-			inline void add(const void* buffer, size_t size){ addBuffer(buffer, size); }
-			void setBuffer(const void* buffer, size_t size);
-			inline void set(const void* buffer, size_t size){ setBuffer(buffer, size); }
+			void addBuffer(const void* buffer, int64 size);
+			inline void add(const void* buffer, int64 size){ addBuffer(buffer, size); }
+			void setBuffer(const void* buffer, int64 size);
+			inline void set(const void* buffer, int64 size){ setBuffer(buffer, size); }
 			//! \brief	删除buffer内容:不更改内存大小。
-			void delBuffer(size_t start, size_t length);
-			inline void del(size_t start, size_t length){ delBuffer(start, length); }
+			void delBuffer(int64 start, int64 length);
+			inline void del(int64 start, int64 length){ delBuffer(start, length); }
 			void clearBuffer();
 			inline void clear(){ clearBuffer(); }
 			//! \brief	扩展内存大小。
 			//! \note	如果已满足扩展大小则不做操作。
 			//! \note	扩展不会影响已有的数据。
-			void reserve(size_t size);
+			void reserve(int64 size);
 			void print(std::string expand = "")const;
-			void setMaxLimit(size_t _maxLimit){ m_MaxLimit = _maxLimit; }
+			void setMaxLimit(int64 _maxLimit){ m_MaxLimit = _maxLimit; }
 
 			inline bool empty()const{ return m_CurBufferSize == 0; }
 			//! \note	直接操作内存.仅用于内容修改。：如有需要可以使用，但请慎重使用。
@@ -71,15 +74,15 @@ namespace Universal{
 			inline void* getBuffer(){ return m_Buffer; }
 			inline void* buffer(){ return m_Buffer; }
 			inline const void* getBuffer()const{ return m_Buffer; }
-			inline size_t getBufferSize()const{ return m_CurBufferSize; }
-			inline size_t curSize()const{ return m_CurBufferSize; }
-			inline size_t maxSize()const{ return m_MaxBufferSize; }
-			inline size_t maxLimit()const{ return m_MaxLimit; }
+			inline int64 getBufferSize()const{ return m_CurBufferSize; }
+			inline int64 curSize()const{ return m_CurBufferSize; }
+			inline int64 maxSize()const{ return m_MaxBufferSize; }
+			inline int64 maxLimit()const{ return m_MaxLimit; }
 		private:
 			void* m_Buffer;
-			size_t m_CurBufferSize;
-			size_t m_MaxBufferSize;
-			size_t m_MaxLimit;
+			int64 m_CurBufferSize;
+			int64 m_MaxBufferSize;
+			int64 m_MaxLimit;
 	};
 	typedef boost::shared_ptr<BinaryMemory> BinaryMemoryPtr;
 }
